Added: Insert method now returns id, added get method in NotesDao, NotesRepository and NotesRepositoryImpl Modified: id type from Int to Long in Note and NoteEntity

+++ app/src/main/java/com/brainbowfx/android/simplenotes/data/database/models/NoteEntity.kt
@@ -6,7 +6,7 @@ import com.brainbowfx.android.simplenotes.data.mappers.Converters
 @Fts4
 @Entity(tableName = "notes")
 class NoteEntity(
-    @PrimaryKey(autoGenerate = true) @ColumnInfo(name = "rowid") var id: Int = 0,
+    @PrimaryKey(autoGenerate = true) @ColumnInfo(name = "rowid") var id: Long = 0,
     val title: String,
     val text: String,
     val date: Long,
